+++ client/src/assets/search.ts
@@ -1,3 +1,4 @@
+import debounce from "lodash/debounce";
 import { ref, watch, type Ref } from "vue";
 
 import type { ApiAsset } from "../apiTypes";
@@ -10,12 +11,15 @@ interface AssetSearch {
     filter: Ref<string>;
     results: Ref<ApiAsset[]>;
     loading: Ref<boolean>;
+    includeSharedAssets: Ref<boolean>;
 }
 
 export function useAssetSearch(searchBar: Ref<HTMLInputElement | null>): AssetSearch {
     const filter = ref("");
     const results = ref<ApiAsset[]>([]);
     const loading = ref(false);
+    const includeSharedAssets = ref(false);
+
     watch(assetState.currentFolder, () => {
         filter.value = "";
     });
@@ -25,20 +29,26 @@ export function useAssetSearch(searchBar: Ref<HTMLInputElement | null>): AssetSe
         searchBar.value?.focus();
     }
 
-    watch(filter, async (filter) => {
-        if (filter.length < 3) {
-            results.value = [];
-            return;
-        }
-
+    async function search(query: string): Promise<void> {
         loading.value = true;
-        const data = (await socket.emitWithAck("Asset.Search", filter)) as ApiAsset[];
+        const data = (await socket.emitWithAck("Asset.Search", query, includeSharedAssets.value)) as ApiAsset[];
         for (const asset of data) {
             assetState.mutableReactive.idMap.set(asset.id, asset);
         }
         results.value = data;
         loading.value = false;
+    }
+
+    const debouncedSearch = debounce(search, 300);
+
+    watch([filter, includeSharedAssets], async ([filter]) => {
+        if (filter.length < 3) {
+            results.value = [];
+            return;
+        }
+
+        await debouncedSearch(filter);
     });
 
-    return { clear, filter, results, loading };
+    return { clear, filter, results, loading, includeSharedAssets };
 }