+++ client/src/game/api/emits/shape/core.ts
@@ -21,7 +21,7 @@ import { wrapSocket, socket } from "../../socket";
 export const sendShapeAdd = wrapSocket<ShapeAdd>("Shape.Add");
 export const sendRemoveShapes = (data: TemporaryShapes): void => {
     if (data.uuids.length === 0) {
-        if (process.env.NODE_ENV === "production") {
+        if (import.meta.env.NODE_ENV === "production") {
             console.error(
                 "Attempted to send shape removal request for 0 shapes. If you think this is a bug, please report this!",
             );