+++ server/src/api/socket/constants.py
@@ -1,3 +1,4 @@
+ADMIN_NS = "/admin"
 ASSET_NS = "/pa_assetmgmt"
 DASHBOARD_NS = "/dashboard"
 GAME_NS = "/planarally"